가게 멤버별 review 조회 api 구현

+++ demo/src/main/java/umc/study/web/dto/ReviewRequest.java
@@ -14,10 +14,8 @@ public class ReviewRequest {
     public static class AddReviewDTO {
 
         @NotNull
-        private float star;
+        private double star;
         private String body;
 
     }
-
-
 }
